Add GPS Network RTK messages to gps.go
